% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip}
\alias{clip}
\title{Cluster wise sign-flipping score test}
\usage{
clip(formula, data, cluster,
n_flips = 5000, alternative= "two.sided", seed = NULL, V = NULL)
}
\arguments{
\item{formula}{A model \code{\link{formula}}. For multivariate outcomes, \code{formula}
may refer to a matrix response in \code{data} (as in \code{\link[stats]{lm}})
or to a column in \code{data}.}

\item{data}{A \code{data.frame} containing the variables referenced in \code{formula}.}

\item{cluster}{Cluster identifier. Character string naming the clustering column in \code{data}.}

\item{n_flips}{Integer. Number of flips to generate. Default is 5000.}

\item{alternative}{Character string specifying the alternative hypothesis used to test the fixed effect
coefficients. One of \code{"two.sided"}, \code{"greater"}, \code{"less"}.
Default is \code{"two.sided"}.}

\item{seed}{Optional integer seed for reproducibility.}

\item{V}{Working covariance matrix to be used inside the score contributions.
It can be a single matrix (univariate case, one response) or a list of matrices (one per outcome).
Default is NULL, i.e., identity matrix.}
}
\value{
A \code{jointest} object, i.e., a list containing the following objects:
\describe{
  \item{Tspace}{\code{data.frame} where rows represents the sign-flipping transformed (plus the identity one) test and columns the variables.}
  \item{summary_table}{\code{data.frame} containing for each model the estimated parameter(s), score(s), std error(s), test(s), partial correlation(s) and p-value(s).}
  \item{mods}{List of \code{flipscores} objects, one for each outcome.}
  \item{call}{The matched call.}
}
}
\description{
This function implements a *cluster wise sign-flipping score-based test* for
one or multiple outcomes under within-cluster dependence (i.e., non-independent observations).
}
\examples{
library(remmm)
df <- data.frame(y = rnorm(100), x = rnorm(100), id = rep(1:20, each = 5))
V  <- diag(100)
out <- clip(y ~ x, data = df, cluster = "id", V = V)
summary(out)


}
\seealso{
\code{\link[flipscores]{flipscores}}
}
\author{
Angela Andreella
}
